Use CSS grid to align recipe cards
