+++ web/src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
     <Navbar/>
-    <div class="container is-fullhd">
+    <div class="container is-fluid">
       <router-view/>
     </div>
     <footer class="footer">
@@ -53,8 +53,6 @@ export default {
   font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
   margin-top: 60px;
 }
 </style>
