+++ web/src/components/MenuSlot.vue
@@ -1,8 +1,6 @@
 <template>
-  <div class="block">
-    <RecipeCard v-if="!isEmpty" v-bind="recipe" @delete-recipe="deleteRecipe"/>
-    <EmptySlot v-else-if="course" :course="course" @pick-recipe="pickRecipe($event)"/>
-  </div>
+  <RecipeCard class="tile is-child" v-if="!isEmpty" v-bind="recipe" @delete-recipe="deleteRecipe"/>
+  <EmptySlot v-else-if="course" :course="course" @pick-recipe="pickRecipe($event)"/>
 </template>
 
 <script>
